requirements.txt
Django>=4.2
mysqlclient>=2.2
python-decouple>=3.8
django-crispy-forms>=2.1
crispy-bootstrap5>=2024.2
Pillow>=10.0
WeasyPrint>=60.0
gunicorn>=21.0
whitenoise>=6.6
dj-database-url>=2.1
psycopg2-binary>=2.9
